chromium/third_party/blink/web_tests/virtual/no-auto-wpt-origin-isolation/external/wpt/origin-isolation/removing-iframes.sub.https-expected.txt

This is a testharness.js-based test.
PASS Before: messageerror event must occur
PASS Before: setting document.domain must not give sync access
PASS parent: originIsolationRestricted must equal true
PASS child1: originIsolationRestricted must equal false
PASS Remove the iframe and insert new ones
PASS Parent to child2: messageerror event must occur
PASS Parent to child2: setting document.domain must not give sync access
PASS Parent to child3: messageerror event must occur
PASS Parent to child3: setting document.domain must not give sync access
FAIL child2 to child3: message event must occur assert_equals: expected "WebAssembly.Module message received" but got "messageerror"
FAIL child2 to child3: setting document.domain must give sync access assert_equals: expected "accessed document successfully" but got "SecurityError"
FAIL child3 to child2: message event must occur assert_equals: expected "WebAssembly.Module message received" but got "messageerror"
FAIL child3 to child2: setting document.domain must give sync access assert_equals: expected "accessed document successfully" but got "SecurityError"
FAIL child2: originIsolationRestricted must equal false assert_equals: expected false but got true
PASS child3: originIsolationRestricted must equal false
Harness: the test ran to completion.
